+++ flows/landing_zone_delete.py
@@ -49,7 +49,8 @@ class Flow(BaseLinearFlow):
                     project_uuid=self.project_uuid,
                     inject={
                         'zone_uuid': self.flow_data['zone_uuid'],
-                        'info_prefix': 'Running asynchronous job failed',
+                        'flow_name': __name__,
+                        'info_prefix': 'Failed to delete landing zone',
                     },
                 )
             )